# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "manypost API",
    "version": "0.0.1",
    "description": "API do manypost — agendamento/publicação multicanal. Humanos usam sessão Clerk; máquinas usam `Authorization: Bearer mp_live_…`. Erros seguem problem+json (RFC 9457): o campo `title` carrega o código estável do erro."
  },
  "servers": [
    {
      "url": "/",
      "description": "mesma origem (proxy do Next)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "clerkAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "sessão Clerk humana — Authorization: Bearer <token>"
      },
      "apiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key mp_live_ para superfícies de máquina — Authorization: Bearer <chave>"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank"
          },
          "title": {
            "type": "string",
            "description": "código estável do erro",
            "example": "common.not_found"
          },
          "status": {
            "type": "integer",
            "example": 404
          },
          "detail": {
            "type": "string",
            "example": "post não encontrado"
          },
          "extra": {
            "description": "campos extras (ex.: issues de validação)"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      },
...
      "PostGroup": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "example": "SCHEDULED"
          },
          "publishAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "publications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Publication"
            }
          }
        },
        "required": [
          "id",
          "state",
          "publishAt",
          "publications"
        ]
      },
...
      "PostGroupDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "example": "SCHEDULED"
          },
          "publishAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "text": {
            "type": "string",
            "description": "texto base do grupo (sem overrides por canal)"
          },
          "publications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicationDetail"
            }
          }
        },
        "required": [
          "id",
          "state",
          "publishAt",
          "text",
          "publications"
        ]
      },
      "ApprovalLink": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "token opaco ≥256 bits — só aparece aqui"
          },
          "url": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "token",
          "url",
          "expiresAt"
        ]
      },
      "ApprovalLinkStatus": {
        "type": [
          "object",
          "null"
        ],
        "properties": {
          "status": {
            "type": "string",
            "example": "PENDING"
          },
          "feedback": {
            "type": [
              "string",
              "null"
            ]
          },
          "approverName": {
            "type": [
              "string",
              "null"
            ]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "resolvedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "status",
          "feedback",
          "approverName",
          "expiresAt",
          "resolvedAt",
          "createdAt"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/posts": {
      "post": {
        "tags": [
          "posts"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Agenda um post (1 grupo → 1 publicação por canal)",
        "description": "Valida texto/mídia por canal e enfileira. `textByChannel` personaliza o texto do post principal por canal; `thread` cria réplicas encadeadas; `requireApproval` nasce DRAFT aguardando aprovação por link.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 10000
                  },
                  "channelIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "minItems": 1,
                    "maxItems": 20
                  },
                  "publishAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "timezone": {
                    "type": "string",
                    "default": "UTC"
                  },
                  "settingsByChannel": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "textByChannel": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 10000
                    }
                  },
                  "mediaIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "maxItems": 10
                  },
                  "thread": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "text": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 10000
                        },
                        "mediaIds": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "maxItems": 10
                        },
                        "delaySec": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 600
                        }
                      },
                      "required": [
                        "text"
                      ]
                    },
                    "maxItems": 24
                  },
                  "requireApproval": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "text",
                  "channelIds",
                  "publishAt"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "grupo agendado (ou DRAFT se requireApproval)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostGroup"
                }
              }
            }
          },
          "400": {
            "description": "requisição fora do contrato (problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "não encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/posts/{groupId}": {
      "get": {
        "tags": [
          "posts"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Detalhe de um grupo de post",
        "description": "Além dos estados, expõe o conteúdo completo (texto base, texto/settings por canal e réplicas de thread) — é o que o composer usa p/ duplicar um post.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "groupId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "grupo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostGroupDetail"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "não encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "posts"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Edita texto, horário e/ou settings por canal (re-agenda com nova versão de job)",
        "description": "Editar `text` sobrescreve o content de TODAS as publicações (overrides resetam). `settingsByChannel` (chave = channelId) faz merge nos settings de publicação do canal e é validado por provider.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "groupId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 10000
                  },
                  "publishAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "settingsByChannel": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "grupo re-agendado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostGroup"
                }
              }
            }
          },
          "400": {
            "description": "requisição fora do contrato (problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "não encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/posts/{groupId}/cancel": {
      "post": {
        "tags": [
          "posts"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Cancela um grupo agendado (job antigo morre por versão)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "groupId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "grupo cancelado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostGroup"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "não encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/posts/{groupId}/retry": {
      "post": {
        "tags": [
          "posts"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Tenta publicar de novo (FAILED/NEEDS_REVIEW → SCHEDULED)",
        "description": "Kanban \"tentar novamente\". Com `channelId` no corpo, retenta só aquele canal.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "groupId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "channelId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "grupo re-agendado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostGroup"
                }
              }
            }
          },
          "400": {
            "description": "requisição fora do contrato (problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "não encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/posts/{groupId}/approval-link": {
      "post": {
        "tags": [
          "posts",
          "approvals"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Cria o link público de aprovação de um rascunho",
        "description": "Só para grupos DRAFT (requireApproval). Criar de novo revoga o link anterior.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "groupId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "expiresInHours": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 720
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "link criado (token só aqui)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalLink"
                }
              }
            }
          },
          "400": {
            "description": "requisição fora do contrato (problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "não encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "posts",
          "approvals"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Status do link de aprovação mais recente (ou null)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "groupId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "status do link (null se nunca houve)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalLinkStatus"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "posts",
          "approvals"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Revoga o link de aprovação pendente",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "groupId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "resultado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "revoked": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "revoked"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}